+++ packages/ariakit-test/src/__utils.ts
@@ -6,13 +6,58 @@ export type TextField = HTMLInputElement | HTMLTextAreaElement;
 
 const preventMouseEvents = new WeakMap<Document, boolean>();
 
-export const isHappyDOM = typeof window !== "undefined" && "happyDOM" in window;
+export function isHappyDOM(
+  win: Window | null | undefined = typeof window !== "undefined"
+    ? window
+    : undefined,
+) {
+  return !!win && "happyDOM" in win;
+}
 
 export const isBrowser =
   typeof navigator !== "undefined" &&
   !navigator.userAgent.includes("jsdom") &&
   typeof window !== "undefined" &&
-  !isHappyDOM;
+  !isHappyDOM();
+
+// happy-dom doesn't implement the legacy `window.event` global, while jsdom and
+// real browsers expose the event currently being dispatched there for the
+// synchronous duration of the dispatch. React 18 reads `window.event` in
+// `getCurrentEventPriority` to classify an update triggered synchronously inside
+// a native event listener: a `click`/`keydown` yields DiscreteEventPriority
+// (sync lane, flushed in a microtask), whereas a missing `window.event` falls
+// back to DefaultEventPriority (default lane, flushed a macrotask later through
+// the scheduler). That later flush lets a microtask-coalesced store batch run
+// before React commits — e.g. a controlled `<Dialog open>` re-applies its stale
+// `open` prop right after `store.hide()`, transiently re-opening the dialog and
+// corrupting focus restoration on outside-click. Mirror jsdom by exposing the
+// dispatched event on `window.event` only while listeners run, then restoring
+// the previous value — or removing the property again when the environment had
+// none — so nested dispatches stay correct and the global isn't left behind. In
+// practice this only changes the React 18 suite: the controlled-dialog focus
+// divergence it fixes doesn't reproduce under React 19.
+export function withWindowEvent<T>(
+  event: Event,
+  run: () => T,
+  win: Window | null | undefined = typeof window !== "undefined"
+    ? window
+    : undefined,
+): T {
+  if (!isHappyDOM(win)) return run();
+  const eventWindow = win as Window & { event?: Event };
+  const had = Object.prototype.hasOwnProperty.call(eventWindow, "event");
+  const previous = eventWindow.event;
+  eventWindow.event = event;
+  try {
+    return run();
+  } finally {
+    if (had) {
+      eventWindow.event = previous;
+    } else {
+      delete eventWindow.event;
+    }
+  }
+}
 
 export async function flushMicrotasks() {
   await Promise.resolve();
